create a new file services.go and move service part codes from cluster.go into services.go
